NF: extract Operation from MigrateUserData
